requirements.txt
asgiref                       3.8.1
Django                        5.1.4
djangorestframework           3.15.2
djangorestframework-simplejwt 5.3.1
pip                           24.0
PyJWT                         2.10.1
sqlparse                      0.5.3
